+++ NPCs/CuteGastropod.cs
@@ -36,7 +36,7 @@ namespace AssortedCrazyThings.NPCs
 			NPC.noGravity = true;
 			AIType = NPCID.FlyingSnake;
 			AnimationType = NPCID.FlyingSnake;
-			NPC.catchItem = (short)ModContent.ItemType<CuteGastropodItem>();
+			NPC.catchItem = ModContent.ItemType<CuteGastropodItem>();
 		}
 
 		public override bool? CanBeHitByItem(Player player, Item item)
