+++ NPCs/JoyousSlime.cs
@@ -38,7 +38,7 @@ namespace AssortedCrazyThings.NPCs
 			NPC.dontTakeDamageFromHostiles = true;
 			NPC.alpha = 175;
 			NPC.color = new Color(169, 141, 255, 100);
-			NPC.catchItem = (short)ModContent.ItemType<JoyousSlimeItem>();
+			NPC.catchItem = ModContent.ItemType<JoyousSlimeItem>();
 		}
 
 		public override bool? CanBeHitByItem(Player player, Item item)
